FP-33 - changed BigDecimal in length into int in Entity and DTO of Film - refactor service from old generated repository - added test film service

+++ backend/src/main/java/io/swagger/repository/FilmDTO.java
@@ -1,6 +1,5 @@
 package io.swagger.repository;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import io.swagger.types.MoodType;
@@ -8,7 +7,7 @@ import io.swagger.types.MoodType;
 public record FilmDTO(
         UUID id,
         String title,
-        BigDecimal lengthInMinutes,
+        int lengthInMinutes,
         MoodType mood,
         String releaseYear) {
 }
